# code/LogicApi/MCWebAPI/Resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Prismarine RESTful Web API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/MinecraftServer/{id}": {
      "get": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Gets the informations of a server. (Auth)",
        "operationId": "GetServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the server.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested server object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedPublic.Model.IMinecraftServer"
                }
              }
            }
          },
          "400": {
            "description": "The server with the specified id does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Deletes a server from the system. (Auth)",
        "operationId": "DeleteServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the server",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The server is deleted. Nothing more."
          },
          "400": {
            "description": "The server with the specified id does not exist or an exception happened during the deletion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Modifies the server information. (Auth)",
        "operationId": "ModifyServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the server to modify",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "new values",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedPublic.DTOs.ModifyServerDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The server is deleted. Nothing more."
          },
          "400": {
            "description": "The server with the specified id does not exist or an exception happened during the deletion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/commands": {
      "post": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Writes a command to the server. (Auth)",
        "operationId": "WriteCommandToServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the minecraft server",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "command data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIModel.DTOs.CommandDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The command is executed."
          },
          "400": {
            "description": "The server with the specified id does not exist or an exception happened during the command execution.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/toggle": {
      "post": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Toggles the minecraft server on and off. (Auth)",
        "operationId": "ToggleServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the minecraft server.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The server is either started or deleted, depending on the state of it."
          },
          "400": {
            "description": "The server with the specified id does not exist or an exception happened during the toggle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/backups": {
      "post": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Backs up the server. (Auth)",
        "operationId": "BackupServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedPublic.DTOs.BackupDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Gets all the backups of a server. (Auth)",
        "operationId": "GetBackups",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SharedPublic.Model.IBackup"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/backups/{backupName}": {
      "delete": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Deletes a backup of a server. (Auth)",
        "operationId": "DeleteBackup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "backupName",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "pattern": "^[\\w\\W]",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIModel.Responses.ExceptionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Restores a backup of a server. (Auth)",
        "operationId": "RestoreBackup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "backupName",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "pattern": "^[\\w\\W]",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/properties": {
      "get": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Gets all the properties of a server (Auth)",
        "operationId": "GetProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Modifies the server properties (Auth)",
        "operationId": "ModifyProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedPublic.DTOs.MinecraftServerPropertiesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/players/{username}": {
      "get": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Gets a player's data on the server. (Auth)",
        "operationId": "GetPlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/MinecraftServer/{id}/icon": {
      "get": {
        "tags": [
          "MinecraftServer"
        ],
        "summary": "Gets the icon of the server. (Auth)",
        "operationId": "GetIcon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIModel.DTOs.CommandDto": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "APIModel.Responses.ExceptionDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "isInternalException": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Application.Minecraft.Versions.IMinecraftVersion": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "version": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "downloadUrl": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "isDownloaded": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.DTOs.BackupDto": {
        "type": "object",
        "properties": {
          "backupName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SharedPublic.DTOs.MinecraftServerPropertiesDto": {
        "type": "object",
        "properties": {
          "allowFlight": {
            "type": "boolean",
            "default": true,
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "default": "normal",
            "nullable": true
          },
          "enableCommandBlock": {
            "type": "boolean",
            "default": true,
            "nullable": true
          },
          "enforceSecureProfile": {
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "gamemode": {
            "type": "string",
            "default": "survival",
            "nullable": true
          },
          "hardcore": {
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "maxWorldSize": {
            "maximum": 29999984,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "default": 29999984,
            "nullable": true
          },
          "motd": {
            "maxLength": 59,
            "type": "string",
            "default": "A Minecraft Server",
            "nullable": true
          },
          "pvp": {
            "type": "boolean",
            "default": true,
            "nullable": true
          },
          "simulationDistance": {
            "maximum": 32,
            "minimum": 3,
            "type": "integer",
            "format": "int32",
            "default": 10,
            "nullable": true
          },
          "spawnMonsters": {
            "type": "boolean",
            "default": true,
            "nullable": true
          },
          "spawnProtection": {
            "maximum": 30000,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "default": 0,
            "nullable": true
          },
          "viewDistance": {
            "maximum": 32,
            "minimum": 3,
            "type": "integer",
            "format": "int32",
            "default": 10,
            "nullable": true
          },
          "whiteList": {
            "type": "boolean",
            "default": false,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.DTOs.ModifyServerDto": {
        "type": "object",
        "properties": {
          "newName": {
            "maxLength": 35,
            "minLength": 4,
            "type": "string",
            "default": null,
            "nullable": true
          },
          "version": {
            "type": "string",
            "default": null,
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.Enums.BackupType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "SharedPublic.Model.IBackup": {
        "type": "object",
        "properties": {
          "serverId": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "type": {
            "$ref": "#/components/schemas/SharedPublic.Enums.BackupType"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.Model.ILogMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "messageType": {
            "$ref": "#/components/schemas/SharedPublic.Model.LogMessageType"
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.Model.IMinecraftServer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "serverName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/SharedPublic.Model.ServerStatus"
          },
          "statusMessage": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "isRunning": {
            "type": "boolean",
            "readOnly": true
          },
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedPublic.Model.ILogMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "onlineFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedPublic.Model.IPlayerSimple"
            },
            "nullable": true,
            "readOnly": true
          },
          "storageSpace": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "storageBytes": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "mcVersion": {
            "$ref": "#/components/schemas/Application.Minecraft.Versions.IMinecraftVersion"
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.Model.IPlayerSimple": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "onlineFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "pastOnlineTicks": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SharedPublic.Model.LogMessageType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "SharedPublic.Model.ServerStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Standard Authorization header using the Bearer scheme (JWT). Example: \"bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}